Move HueConfiguration to application
