+++ Guppi/Properties/launchSettings.json
@@ -2,7 +2,7 @@
   "profiles": {
     "Guppi.Console": {
       "commandName": "Project",
-      "commandLineArgs": "notes"
+      "commandLineArgs": "lights on"
     }
   }
 }
